[AIP-21] Move operators.hive_to_druid providers.apache.druid.operators.hive_to_druid
